[feat] 구독 시 플랜 정보 조회하여 알맞은 코인 추가하도록 변경

+++ backend/src/main/java/com/sikgu/sikgubackend/entity/User.java
@@ -74,7 +74,7 @@ public class User extends BaseEntity {
         this.carts.add(cart);
     }
 
-    public void addCoins(int amount) {
+    public void addCoins(long amount) {
         if (this.coins == null) {
             this.coins = 0L;
         }
